\name{ExpectedForce-package}
\alias{ExpectedForce-package}
\alias{ExpectedForce}
\docType{package}
\title{
\packageTitle{ExpectedForce}
}
\description{
\packageDescription{ExpectedForce}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ExpectedForce}
\packageIndices{ExpectedForce}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{ExpectedForce}

Maintainer: \packageMaintainer{ExpectedForce}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
